Move load_tts_library() to RHVoice.py
