+++ src/nvda-synthDriver/__init__.py
@@ -37,6 +37,7 @@ from RHVoice import RHVoice_tts_engine, RHVoice_message, RHVoice_lib_path
 from RHVoice import RHVoice_init_params, RHVoice_callback_types, RHVoice_callbacks
 from RHVoice import RHVoice_synth_params
 from RHVoice import RHVoice_voice_info
+from RHVoice import load_tts_library
 
 import config
 import globalVars
@@ -120,34 +121,6 @@ class RHVoice_capitals_mode:
 
 class RHVoice_synth_flag:
 	dont_clip_rate=1
-
-def load_tts_library():
-	try:
-		lib=ctypes.CDLL(RHVoice_lib_path.encode("mbcs"))
-	except TypeError:
-		lib=ctypes.CDLL(RHVoice_lib_path)
-	lib.RHVoice_get_version.restype=c_char_p
-	lib.RHVoice_new_tts_engine.argtypes=(POINTER(RHVoice_init_params),)
-	lib.RHVoice_new_tts_engine.restype=RHVoice_tts_engine
-	lib.RHVoice_delete_tts_engine.argtypes=(RHVoice_tts_engine,)
-	lib.RHVoice_delete_tts_engine.restype=None
-	lib.RHVoice_get_number_of_voices.argtypes=(RHVoice_tts_engine,)
-	lib.RHVoice_get_number_of_voices.restype=c_uint
-	lib.RHVoice_get_voices.argtypes=(RHVoice_tts_engine,)
-	lib.RHVoice_get_voices.restype=POINTER(RHVoice_voice_info)
-	lib.RHVoice_get_number_of_voice_profiles.argtypes=(RHVoice_tts_engine,)
-	lib.RHVoice_get_number_of_voice_profiles.restype=c_uint
-	lib.RHVoice_get_voice_profiles.argtypes=(RHVoice_tts_engine,)
-	lib.RHVoice_get_voice_profiles.restype=POINTER(c_char_p)
-	lib.RHVoice_are_languages_compatible.argtypes=(RHVoice_tts_engine,c_char_p,c_char_p)
-	lib.RHVoice_are_languages_compatible.restype=c_int
-	lib.RHVoice_new_message.argtypes=(RHVoice_tts_engine,c_char_p,c_uint,c_int,POINTER(RHVoice_synth_params),c_void_p)
-	lib.RHVoice_new_message.restype=RHVoice_message
-	lib.RHVoice_delete_message.arg_types=(RHVoice_message,)
-	lib.RHVoice_delete_message.restype=None
-	lib.RHVoice_speak.argtypes=(RHVoice_message,)
-	lib.RHVoice_speak.restype=c_int
-	return lib
 
 def escape_text(text):
 	parts=list()
